+++ assets_test.go
@@ -6,9 +6,15 @@ import (
 	"github.com/thrisp/flotilla/resources"
 )
 
-var TestAsset *AssetFS = &AssetFS{resources.Asset, resources.AssetDir, resources.AssetNames, ""}
+var TestAsset AssetFS = NewAssetFS(
+	resources.Asset,
+	resources.AssetDir,
+	resources.AssetNames,
+	"",
+)
 
 func TestAssets(t *testing.T) {
+	//spew.Dump(TestAsset)
 	//f, _ := resources.Asset("/assets/templates/test_asset.html")
 	//fd, _ := resources.AssetDir("assets")
 	//names := resources.AssetNames()