+++ configuration_test.go
@@ -12,9 +12,9 @@ func FauxConf() Configuration {
 }
 
 func TestConfiguration(t *testing.T) {
-	a := New(
-		"conf",
-		Mode("testing", true),
+	a := testApp(
+		t,
+		"configurationTest",
 		Mode("prodnnuction", true),
 		EnvItem("value:set", "section_value:set"),
 		FauxConf(),
@@ -26,37 +26,32 @@ func TestConfiguration(t *testing.T) {
 
 	var non bool
 
-	conffunc := func(c Ctx) {
-		tm, _ := c.Call("mode")
-		testmodes = tm.(*Modes)
-		envval, _ = CheckStore(c, "VALUE")
-		envsectionval, _ = CheckStore(c, "SECTION_VALUE")
-		_, non = CheckStore(c, "NON")
+	confTester := func(t *testing.T) Manage {
+		return func(c Ctx) {
+			tm, _ := c.Call("mode")
+			testmodes = tm.(*Modes)
+			if !testmodes.Development || !testmodes.Testing || testmodes.Production {
+				t.Errorf("Modes not properly set: %+v", testmodes)
+			}
+			envval, _ = CheckStore(c, "VALUE")
+			if envval.Value != "set" {
+				t.Errorf(`EnvItem was not set properly; was EnvItem("value:set"), but retrieved Store value is %s`, envval)
+			}
+			envsectionval, _ = CheckStore(c, "SECTION_VALUE")
+			if envsectionval.Value != "set" {
+				t.Errorf(`EnvItem was not set properly; was EnvItem("section_value:set"), but retrieved Store value is %s`, envsectionval)
+			}
+			_, non = CheckStore(c, "NON")
+			if non {
+				t.Errorf(`A value was found in the store that should not exist.`)
+			}
+
+		}
 	}
 
-	a.GET("/configuration", conffunc)
+	exp, _ := NewExpectation(200, "GET", "/configuration", confTester)
 
-	a.Configure(a.Configuration...)
-
-	p := NewPerformer(t, a, 200, "GET", "/configuration")
-
-	performFor(p)
-
-	if !testmodes.Development || !testmodes.Testing || testmodes.Production {
-		t.Errorf("Modes not properly set: %+v", testmodes)
-	}
-
-	if envval.Value != "set" {
-		t.Errorf(`EnvItem was not set properly; was EnvItem("value:set"), but retrieved Store value is %s`, envval)
-	}
-
-	if envsectionval.Value != "set" {
-		t.Errorf(`EnvItem was not set properly; was EnvItem("section_value:set"), but retrieved Store value is %s`, envsectionval)
-	}
-
-	if non {
-		t.Errorf(`A value was found in the store that should not exist.`)
-	}
+	SimplePerformer(t, a, exp).Perform()
 
 	if !fauxconf {
 		t.Errorf(`Arbitrary Configuration function FauxConf was not properly set or used.`)