+++ extensions.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 	"reflect"
 
+	"github.com/thrisp/flotilla/engine"
 	"github.com/thrisp/flotilla/session"
 	"github.com/thrisp/flotilla/xrr"
 )
@@ -123,7 +124,8 @@ func iswritten(c *ctx) bool {
 
 func redirect(c *ctx, code int, location string) error {
 	if code >= 300 && code <= 308 {
-		c.push(func(pc Ctx) {
+		c.bounce(func(pc Ctx) {
+			releasesession(c)
 			http.Redirect(c.RW, c.Request, location, code)
 			c.RW.WriteHeaderNow()
 		})
@@ -185,6 +187,7 @@ func returnsession(c *ctx) session.SessionStore {
 func releasesession(c *ctx) error {
 	if c.Session != nil {
 		if !c.RW.Written() {
+			c.Flasher.Out(c.Session)
 			c.Session.SessionRelease(c.RW)
 		}
 	}
@@ -205,56 +208,88 @@ func startsession(c *ctx, s *session.Manager) error {
 }
 
 var flashfxtension = map[string]interface{}{
+	"flasher": flshr,
 	"flash":   flash,
-	"flashes": flashes,
-	"flashed": flashed,
 }
 
 var FlashFxtension Fxtension = MakeFxtension("flashfxtension", flashfxtension)
 
-func flash(c *ctx, category string, message string) error {
-	if fl := c.Session.Get("_flashes"); fl != nil {
-		if fls, ok := fl.(map[string][]string); ok {
-			fls[category] = append(fls[category], message)
-			c.Session.Set("_flashes", fls)
-		}
-	} else {
-		fl := make(map[string][]string)
-		fl[category] = append(fl[category], message)
-		c.Session.Set("_flashes", fl)
+type Flashes map[string][]string
+
+type Flasher interface {
+	Write(string) []string
+	WriteAll() Flashes
+	In(session.SessionStore) bool
+	Out(session.SessionStore) bool
+	Flash(string, string)
+}
+
+func NewFlasher() Flasher {
+	return &flasher{}
+}
+
+type flasher struct {
+	readOnce bool
+	flashes  Flashes
+}
+
+func (f *flasher) Write(key string) []string {
+	if ret, ok := f.flashes[key]; ok {
+		f.readOnce = true
+		return ret
 	}
 	return nil
 }
 
-func flashes(c *ctx, categories []string) []string {
-	var ret []string
-	if fl := c.Session.Get("_flashes"); fl != nil {
-		if fls, ok := fl.(map[string][]string); ok {
-			for k, v := range fls {
-				if existsIn(k, categories) {
-					ret = v
-					delete(fls, k)
-				}
-			}
-			c.Session.Set("_flashes", fls)
-		}
+func (f *flasher) WriteAll() Flashes {
+	ret := make(Flashes)
+	for k, v := range f.flashes {
+		ret[k] = v
 	}
+	f.readOnce = true
+	f.flashes = nil
 	return ret
 }
 
-func flashed(c *ctx) map[string][]string {
-	var ret map[string][]string
-	if fl := c.Session.Get("_flashes"); fl != nil {
-		if fls, ok := fl.(map[string][]string); ok {
-			ret = fls
+func (f *flasher) In(s session.SessionStore) bool {
+	if in := s.Get("_flashes"); in != nil {
+		if inf, ok := in.(Flashes); ok {
+			f.flashes = inf
+			return true
 		}
 	}
-	c.Session.Delete("_flashes")
-	return ret
+	return false
 }
 
-// MakeCtxFxtension creates an Fxtension with miscellaneous functions for the
-// provided App.
+func (f *flasher) Out(s session.SessionStore) bool {
+	if err := s.Set("_flashes", f.flashes); err != nil {
+		return true
+	}
+	return false
+}
+
+func (f *flasher) Flash(key, value string) {
+	if f.flashes == nil {
+		f.flashes = make(Flashes)
+	}
+	f.flashes[key] = append(f.flashes[key], value)
+}
+
+func Flshr(c Ctx) Flasher {
+	fl, _ := c.Call("flasher")
+	return fl.(Flasher)
+}
+
+func flshr(c *ctx) (Flasher, error) {
+	return c.Flasher, nil
+}
+
+func flash(c *ctx, category, value string) error {
+	c.Flasher.Flash(category, value)
+	return nil
+}
+
+// MakeCtxFxtension creates a utility Fxtension with miscellaneous functions.
 func MakeCtxFxtension(a *App) Fxtension {
 	ctxfxtension := map[string]interface{}{
 		"env":            envqueryfunc(a),
@@ -262,8 +297,11 @@ func MakeCtxFxtension(a *App) Fxtension {
 		"get":            getdata,
 		"mode":           currentmodefunc(a),
 		"out":            out(a),
+		"emit":           emit(a),
 		"panics":         panics,
 		"panicsignal":    panicsignalfunc(a),
+		"params":         currentparams,
+		"paramString":    paramString,
 		"push":           push,
 		"rendertemplate": rendertemplatefunc(a),
 		"request":        currentrequest,
@@ -314,6 +352,10 @@ func currentmodefunc(a *App) func(c *ctx) *Modes {
 	}
 }
 
+func currentparams(c *ctx) engine.Params {
+	return c.Params
+}
+
 func panics(c *ctx) xrr.ErrorMsgs {
 	return c.Result.Errors().ByType(xrr.ErrorTypePanic)
 }
@@ -325,6 +367,15 @@ func panicsignalfunc(a *App) func(*ctx, string) error {
 	}
 }
 
+func paramString(c *ctx, key string) string {
+	for _, v := range c.Params {
+		if v.Key == key {
+			return v.Value
+		}
+	}
+	return ""
+}
+
 func out(a *App) func(*ctx, string) error {
 	return func(c *ctx, msg string) error {
 		a.Messaging.Out(msg)
@@ -332,9 +383,21 @@ func out(a *App) func(*ctx, string) error {
 	}
 }
 
+func emit(a *App) func(*ctx, string) error {
+	return func(c *ctx, msg string) error {
+		a.Messaging.Emit(msg)
+		return nil
+	}
+}
+
 func push(c *ctx, m Manage) error {
 	c.push(m)
 	return nil
+}
+
+func CurrentRequest(c Ctx) *http.Request {
+	req, _ := c.Call("request")
+	return req.(*http.Request)
 }
 
 func currentrequest(c *ctx) *http.Request {