+++ flotilla_test.go
@@ -1,73 +1,38 @@
 package flotilla
 
-import (
-	"math/rand"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-)
+import "testing"
 
-var METHODS []string = []string{"GET", "POST", "PATCH", "DELETE", "PUT", "OPTIONS", "HEAD"}
-
-func testApp(name string, routes ...*Route) *App {
-	f := New(name, Mode("testing", true))
-	for _, r := range routes {
-		f.Manage(r)
+func testApp(t *testing.T, name string, conf ...Configuration) *App {
+	conf = append(conf, Mode("testing", true))
+	a := New(name, conf...)
+	mkTestQueues(t, a)
+	err := a.Configure()
+	if err != nil {
+		t.Errorf("Error in app configuration: %s", err.Error())
 	}
-	f.Configure()
-	return f
+	return a
 }
 
 func TestSimple(t *testing.T) {
-	a := testApp("simple")
+	a := testApp(t, "simple")
 	if a.Name() != "simple" {
 		t.Errorf(`App name was %s, expected "simple"`, a.Name())
 	}
 }
 
-type Performer struct {
-	t        *testing.T
-	h        http.Handler
-	code     int
-	request  *http.Request
-	response *httptest.ResponseRecorder
-}
-
-func NewPerformer(t *testing.T, a *App, code int, method, path string) *Performer {
-	req, _ := http.NewRequest(method, path, nil)
-	res := httptest.NewRecorder()
-
-	return &Performer{
-		t:        t,
-		h:        a,
-		code:     code,
-		request:  req,
-		response: res,
-	}
-}
-
-func performFor(p *Performer) *Performer {
-	p.h.ServeHTTP(p.response, p.request)
-
-	if p.response.Code != p.code {
-		p.t.Errorf("%s :: %s\nStatus code should be %d, was %d\n", p.request.Method, p.request.URL.Path, p.code, p.response.Code)
-	}
-
-	return p
-}
-
 func testRouteOK(method string, t *testing.T) {
 	var passed bool = false
 
-	r := NewRoute(method, "/test", false, []Manage{func(c Ctx) {
-		passed = true
-	}})
+	exp, _ := NewExpectation(
+		200,
+		method,
+		"/test",
+		func(t *testing.T) Manage { return func(c Ctx) { passed = true } },
+	)
 
-	f := testApp("flotilla_testRouteOK", r)
+	app := testApp(t, "flotilla_testRouteOK")
 
-	p := NewPerformer(t, f, 200, method, "/test")
-
-	performFor(p)
+	SimplePerformer(t, app, exp).Perform()
 
 	if passed == false {
 		t.Errorf("Route handler %s was not invoked.", method)
@@ -80,66 +45,67 @@ func TestRouteOK(t *testing.T) {
 	}
 }
 
-type rx struct {
+type tx struct {
 	method  string
 	passed  bool
 	runonce bool
-	rt      *Route
+}
+
+func txs() []*tx {
+	var ret []*tx
+	for _, m := range METHODS {
+		ret = append(ret, &tx{method: m})
+	}
+	return ret
+}
+
+func rmanage(t *testing.T, x *tx) Manage {
+	return func(c Ctx) {
+		req := CurrentRequest(c)
+		if req.Method == x.method {
+			x.passed = true
+			x.runonce = !x.runonce
+		}
+		if req.Method != x.method {
+			t.Errorf("Request was %s, but Manager expected %s", req.Method, x.method)
+		}
+	}
 }
 
 func TestMultipleRoutesSameMethodOK(t *testing.T) {
-	var rtx []*rx
-	for _, m := range METHODS {
-		mkrx := &rx{
-			method:  m,
-			runonce: false,
-			passed:  false,
-		}
-		mkrx.rt = NewRoute(m, "/test", false, []Manage{func(c Ctx) { mkrx.passed, mkrx.runonce = true, true }})
-		rtx = append(rtx, mkrx)
+	var x []Expectation
+	ctxs := txs()
+	for _, r := range ctxs {
+		m := rmanage(t, r)
+		nx, _ := NewExpectation(
+			200,
+			r.method,
+			"/test",
+			func(t *testing.T) Manage {
+				return m
+			},
+		)
+		x = append(x, nx)
 	}
-	var rts []*Route
-	for _, x := range rtx {
-		rts = append(rts, x.rt)
-	}
-	a := testApp("testRoutesOK", rts...)
-	for _, m := range METHODS {
-		p := NewPerformer(t, a, 200, m, "/test")
-		performFor(p)
-	}
-	for _, x := range rtx {
-		if x.passed != true && x.runonce != true {
-			t.Errorf("Route with same path, but differing method was not registered or run: %+v", x)
+	app := testApp(t, "multipleRoutesSameMethodOk")
+	MultiPerformer(t, app, x...).Perform()
+	for _, ctx := range ctxs {
+		if !ctx.passed && !ctx.runonce {
+			t.Errorf("Multiple route same method error: %+v", ctx)
 		}
 	}
-}
-
-func methodNotMethod(method string) string {
-	newmethod := METHODS[rand.Intn(len(METHODS))]
-	if newmethod == method {
-		methodNotMethod(newmethod)
-	}
-	return newmethod
 }
 
 func testRouteNotOK(method string, t *testing.T) {
-	var passed bool = false
+	exp, _ := NotFoundExpectation(
+		method,
+		"/test",
+		func(t *testing.T) Manage { return func(c Ctx) {} },
+	)
 
-	othermethod := methodNotMethod(method)
+	app := testApp(t, "flotilla_testRouteNotOk")
 
-	r := NewRoute(othermethod, "/test_notfound", false, []Manage{func(c Ctx) {
-		passed = true
-	}})
-
-	f := testApp("flotilla_testRouteNotOk", r)
-
-	p := NewPerformer(t, f, 404, method, "/test_notfound")
-
-	performFor(p)
-
-	if passed == true {
-		t.Errorf("Route handler %s was not invoked.", method)
-	}
+	SimplePerformer(t, app, exp).Perform()
 }
 
 func TestRouteNotOK(t *testing.T) {