+++ httpstatus_test.go
@@ -3,6 +3,7 @@ package flotilla
 import (
 	"bytes"
 	"fmt"
+	"net/http/httptest"
 	"strings"
 	"testing"
 )
@@ -13,45 +14,58 @@ func callStatus(status int) Manage {
 	}
 }
 
-func callStatusRoute(method, route string, status int) *Route {
-	return NewRoute(method, route, false, []Manage{callStatus(status)})
-}
-
 func routestring(status int) string {
 	return fmt.Sprintf("/call%d", status)
 }
 
-func testStatus(method, expects string, status int, t *testing.T) {
-	a := testApp("statuses", callStatusRoute(method, routestring(status), status))
+func testStatus(t *testing.T, status int, method, expects string) {
+	a := testApp(t, "statuses")
 
-	p := NewPerformer(t, a, status, method, routestring(status))
+	exp, _ := NewExpectation(
+		status,
+		method,
+		routestring(status),
+		func(t *testing.T) Manage {
+			return callStatus(status)
+		},
+	)
+	exp.SetPost(
+		func(t *testing.T, r *httptest.ResponseRecorder) {
+			if bytes.Compare(r.Body.Bytes(), []byte(expects)) != 0 {
+				t.Errorf("Status test expected %s, but got %s.", expects, r.Body.String())
+			}
+		},
+	)
 
-	performFor(p)
-
-	if bytes.Compare(p.response.Body.Bytes(), []byte(expects)) != 0 {
-		t.Errorf("Status test expected %s, but got %s.", expects, p.response.Body.String())
-	}
+	SimplePerformer(t, a, exp).Perform()
 }
 
 func TestStatus(t *testing.T) {
 	for _, m := range METHODS {
-		testStatus(m, "418 I'm a teapot", 418, t)
+		testStatus(t, 418, m, "418 I'm a teapot")
 	}
-}
-
-func testpanic(c Ctx) {
-	panic("Test panic!")
 }
 
 func test500(method string, t *testing.T) {
-	pnc := NewRoute(method, routestring(500), false, []Manage{testpanic})
-	a := testApp("panic", pnc)
-	p := NewPerformer(t, a, 500, method, routestring(500))
-	performFor(p)
-	if !strings.Contains(p.response.Body.String(), "Test panic!") {
-		t.Errorf(`Status test 500 expected to contain "Test Panic!", but did not.`)
-	}
-	//fmt.Printf("%+v\n", p.response.Body.String())
+	a := testApp(t, "panic")
+	exp, _ := NewExpectation(
+		500,
+		method,
+		routestring(500),
+		func(t *testing.T) Manage {
+			return func(c Ctx) {
+				panic("Test panic!")
+			}
+		},
+	)
+	exp.SetPost(
+		func(t *testing.T, r *httptest.ResponseRecorder) {
+			if !strings.Contains(r.Body.String(), "Test panic!") {
+				t.Errorf(`Status test 500 expected to contain "Test Panic!", but did not.`)
+			}
+		},
+	)
+	SimplePerformer(t, a, exp).Perform()
 }
 
 func Test500(t *testing.T) {
@@ -64,17 +78,15 @@ func Custom404(c Ctx) {
 	c.Call("serveplain", 404, "I AM NOT FOUND :: 404")
 }
 
-func customStatus404(method, expects string, t *testing.T) {
-	a := testApp("customstatus404")
-
+func customStatus404(t *testing.T, method, expects string) {
+	a := testApp(t, "customstatus404")
 	a.STATUS(404, Custom404)
 
-	p := NewPerformer(t, a, 404, method, "/notfound")
+	z := ZeroExpectationPerformer(t, a, 404, method, "/notfound")
+	z.Perform()
 
-	performFor(p)
-
-	if bytes.Compare(p.response.Body.Bytes(), []byte(expects)) != 0 {
-		t.Errorf("Custom status 404 test expected %s, but got %s.", expects, p.response.Body.String())
+	if bytes.Compare(z.response.Body.Bytes(), []byte(expects)) != 0 {
+		t.Errorf("Custom status 404 test expected %s, but got %s.", expects, z.response.Body.String())
 	}
 }
 
@@ -82,23 +94,34 @@ func Custom418(c Ctx) {
 	c.Call("serveplain", 418, "I AM TEAPOT :: 418")
 }
 
-func customStatus(method, expects string, status int, m Manage, t *testing.T) {
-	a := testApp("customstatus", callStatusRoute(method, routestring(status), status))
-
+func customStatus(t *testing.T, method string, expects string, status int, m Manage) {
+	a := testApp(t, "customstatus")
 	a.STATUS(status, m)
 
-	p := NewPerformer(t, a, status, method, routestring(status))
+	exp, _ := NewExpectation(
+		418,
+		"GET",
+		routestring(status),
+		func(t *testing.T) Manage {
+			return func(c Ctx) {
+				_, _ = c.Call("status", status)
+			}
+		},
+	)
+	exp.SetPost(
+		func(t *testing.T, r *httptest.ResponseRecorder) {
+			if bytes.Compare(r.Body.Bytes(), []byte(expects)) != 0 {
+				t.Errorf("Custom status %d test expected %s, but got %s.", status, expects, r.Body.String())
+			}
+		},
+	)
 
-	performFor(p)
-
-	if bytes.Compare(p.response.Body.Bytes(), []byte(expects)) != 0 {
-		t.Errorf("Custom status %d test expected %s, but got %s.", status, expects, p.response.Body.String())
-	}
+	SimplePerformer(t, a, exp).Perform()
 }
 
 func TestCustomStatus(t *testing.T) {
 	for _, m := range METHODS {
-		customStatus404(m, "I AM NOT FOUND :: 404", t)
-		customStatus(m, "I AM TEAPOT :: 418", 418, Custom418, t)
+		customStatus404(t, m, "I AM NOT FOUND :: 404")
+		customStatus(t, m, "I AM TEAPOT :: 418", 418, Custom418)
 	}
 }