+++ messaging.go
@@ -6,26 +6,23 @@ import (
 	"os"
 )
 
-type (
-	// A byte signal for App messaging.
-	Signal []byte
+// A byte signal for App messaging.
+type Signal []byte
 
-	// A Signal channel for App messaging.
+// A Signal channel for App messaging.
+type Signals chan Signal
+
+// A function taking any string.
+type Queue func(string)
+
+// Messaging encapsulates signalling & logging in an App.
+type Messaging struct {
 	Signals chan Signal
+	Queues  map[string]Queue
+	Logger  *log.Logger
+}
 
-	Queue func(string)
-
-	// Messaging encapsulates signalling & logging in an App.
-	Messaging struct {
-		Signals chan Signal
-		Queues  map[string]Queue
-		Logger  *log.Logger
-	}
-)
-
-var (
-	FlotillaPanic = []byte("flotilla-panic")
-)
+var FlotillaPanic = []byte("flotilla-panic")
 
 func newMessaging() *Messaging {
 	m := &Messaging{}
@@ -35,17 +32,27 @@ func newMessaging() *Messaging {
 	return m
 }
 
-// Out sends the provided string to messaging logger.
+func flush(c Signals, m Signal) {
+	c <- m
+}
+
 func (m *Messaging) Out(message string) {
+	m.Send("out", message)
+}
+
+// Out sends the provided string to messaging logger.
+func (m *Messaging) DefaultOut(message string) {
 	m.Logger.Printf(" %s", message)
 }
 
 // Panic immediately logs the provided string, and sends a FlotillaPanic signal
 // and the message to messaging Signals.
 func (m *Messaging) Panic(message string) {
+	m.Send("panic", message)
+}
+
+func (m *Messaging) DefaultPanic(message string) {
 	log.Println(fmt.Errorf("[Flotilla Panic] %s", message))
-	m.Signals <- FlotillaPanic
-	m.Signals <- []byte(message)
 }
 
 // Emit send the provided message as a Signal to messaging Signals channel.
@@ -56,17 +63,19 @@ func (m *Messaging) Emit(message string) {
 // Send sends the message to the provided queue, with a fall through to Emit if
 // the queue does not exist.
 func (m *Messaging) Send(queue string, message string) {
-	if q, ok := m.Queues[queue]; ok {
-		q(message)
-		return
+	q, ok := m.Queues[queue]
+	if ok {
+		go q(message)
 	}
-	m.Emit(message)
+	if !ok {
+		go m.Emit(message)
+	}
 }
 
 func (m Messaging) defaultqueues() map[string]Queue {
-	q := make(map[string]Queue)
-	q["out"] = m.Out
-	q["panic"] = m.Panic
-	q["emit"] = m.Emit
-	return q
+	return map[string]Queue{
+		"out":   m.DefaultOut,
+		"panic": m.DefaultPanic,
+		"emit":  m.Emit,
+	}
 }