+++ messaging_test.go
@@ -5,6 +5,22 @@ import (
 	"testing"
 )
 
+func testout(t *testing.T, a *App) Queue {
+	return func(message string) {}
+}
+
+func testpanicq(t *testing.T, a *App) Queue {
+	return func(message string) {
+		a.Signals <- FlotillaPanic
+		a.Signals <- []byte(message)
+	}
+}
+
+func mkTestQueues(t *testing.T, a *App) {
+	a.Messaging.Queues["out"] = testout(t, a)
+	a.Messaging.Queues["panic"] = testpanicq(t, a)
+}
+
 func testsignalq(t *testing.T, a *App, against func(*testing.T, Signal)) {
 	go func() {
 		for msg := range a.Signals {
@@ -14,7 +30,7 @@ func testsignalq(t *testing.T, a *App, against func(*testing.T, Signal)) {
 }
 
 func testSignal(method string, t *testing.T) {
-	a := New("testSignal")
+	a := testApp(t, "testSignal", WithQueue("none", func(string) {}))
 	testsignalq(t, a, func(t *testing.T, msg Signal) {
 		m := fmt.Sprintf("%s", msg)
 		if m != "TEST" {
@@ -27,17 +43,14 @@ func testSignal(method string, t *testing.T) {
 		}
 	}
 	a.Queues["testqueue"] = testqueue
-	a.Manage(NewRoute(method, "/test_signal_sent", false, []Manage{func(c Ctx) {
+	a.Manage(NewRoute(defaultRouteConf(method, "/test_signal_sent", []Manage{func(c Ctx) {
 		a.Emit("TEST")
 		for i := 0; i < 10; i++ {
 			a.Send("testqueue", "SENT")
 			a.Send("notaqueue", "TEST")
 		}
-	}}))
-	a.Configure()
-	p := NewPerformer(t, a, 200, method, "/test_signal_sent")
-	performFor(p)
-
+	}})))
+	ZeroExpectationPerformer(t, a, 200, method, "/test_signal_sent").Perform()
 }
 
 func TestSignal(t *testing.T) {